datalink-messenger/requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
tzdata==2024.1
djangorestframework==3.15.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
requests==2.31.0
urllib3==2.2.1